chore: move amount to root
