+++ pactus/transaction/transaction.py
@@ -1,7 +1,7 @@
 from pactus.crypto.address import Address
 from pactus.crypto.private_key import PrivateKey
 from pactus.encoding import encoding
-from pactus.types.amount import Amount
+from pactus.amount import Amount
 
 from ._payload import (
     BondPayload,
